backend/ml/utils.py
import requests
from bs4 import BeautifulSoup
import os

def scrape_url(url):
    """
    Extracts main text content from a given news URL.
    """
    try:
        header = {'User-Agent': 'Mozilla/5.0 (Windows NT 10.0; Win64; x64) AppleWebKit/537.36 (KHTML, like Gecko) Chrome/91.0.4472.124 Safari/537.36'}
        response = requests.get(url, headers=header, timeout=10)
        soup = BeautifulSoup(response.content, 'html.parser')
        
        # Remove script and style elements
        for script in soup(["script", "style"]):
            script.decompose()

        # Get text
        text = soup.get_text()
        # Break into lines and remove leading and trailing whitespace on each
        lines = (line.strip() for line in text.splitlines())
        # Break multi-headlines into a line each
        chunks = (phrase.strip() for line in lines for phrase in line.split("  "))
        # Drop blank lines
        text = '\n'.join(chunk for chunk in chunks if chunk)
        
        # Return first 2000 characters for analysis
        return text[:2000]
    except Exception as e:
        print(f"Scraping error: {e}")
        return ""

def audio_to_text(audio_file):
    """
    Simulated Speech-to-Text logic. 
    In a production app, use OpenAI Whisper or Google Speech-to-Text.
    """
    # For demonstration without heavy binary dependencies
    return "This is a simulated transcription of the audio content regarding the latest news."

def analyze_image_meta(image_file):
    """
    Simulated Image Metadata/Source consistency analysis.
    """
    return {
        "metadata_found": True,
        "edit_software": "None detected",
        "consistency_score": 85
    }
