backend/requirements.txt
django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.3.1
scikit-learn==1.3.2
pandas==2.1.3
joblib==1.3.2
beautifulsoup4==4.12.2
requests==2.31.0
numpy==1.26.2
gunicorn==21.2.0
whitenoise==6.6.0
python-dotenv==1.0.0
